Adding authentication to navigation flow
Adding authentication check before displaying regular navigation.
If a user is not authenticated (and we have not overridden auth for
local development), they will be redirected to the Auth0 sign in page.
When redirected back, we will show the normal app experience.
Adding dependency on native-base to get access to stock UI components.